Starting to set up basic layout
Have a form which takes in two text values and parses them to doubles. Need to also have a dropdown for which action you are doing. Need to store data into database using sqflite. Have this entire layout called from a homepage which will have a view history, add exercise, and add cardio options. This will be the add exercise page.