Added dropdown, crashing occasionally
Added a dropdown, but the value reference seems to be causing the app to crash on boot, not super sure why. Need to edit layout a bit, currently doing a container with a column in it, might move towards more of a fitted box or something. I mostly just want the text input fields to be bigger (decorators?) and to have some vertial padding, with the button lower and bigger also. Also need to add pages for exercise saving and graphs, and of course the sqflite stuff.